streamline README
I've used the style I use in my other READMEs (feedback more than
welcome).  I suspect it's not perfect but gives us a route for using the
README to answer quick questions such as:

* What is this?
* How do I start doing things now with as little effort as possible?
* Where would you like feedback (e.g. issues)?
* How can I learn more about the stack?

The last is handled through links and the others should be through
content.  For a more detailed explanation of gettings started to
contribute I've put things in the CONTRIBUTING.md file.

The reference section on the environment I've simply set in the `.env`
file.  We can discuss why this is or isn't a great idea if it shouldn't
live there.

More pull requests will be coming with other changes as necessary.